[hw 8: архитектура] переключение страниц
