+++ src/pug/events.pug
@@ -1,0 +1,10 @@
+include _panel
+
+h1 Лента событий
+
+.layout
+  for item in events
+    +panel(item)(class=[
+      'layout__item',
+      `layout__item_${item.size}`
+    ])
