+++ src/styl/blocks/nav.styl
@@ -11,7 +11,7 @@
     &:last-child
       margin-right 0
 
-    &_active,
+    &.active,
     body:not(.is-touch-device) &:hover
       color $textColor
