+++ tslint.json
@@ -4,7 +4,8 @@
   "jsRules": {},
   "rules": {
     "max-line-length": [true, 150],
-    "no-magic-numbers": false
+    "no-magic-numbers": false,
+    "no-console": [true, "log"]
   },
   "rulesDirectory": []
 }
